Make NodeSet fully serializable
Currently, the NodeSet is only serializable in one direction so we can
store a subset of information in the build's parameters.

This change makes the NodeSet fully serializable, so we are able to
store it in ZooKeeper and restore it on the executor server.

This is a preparation step to allow locking/unlocking the nodes on the
executor server.

Change-Id: I83c910c0190d8fd1ea17e29f82605508f9c6e2ae