Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-2
> 
> We explicitly chose not to make the tests part of the Zuul deliverable.  We're not going to guarantee an external-facing API, and due to all the test fixtures, they are *huge*.  I think we need a different approach to solve this problem.

namespace packages could be an alternative but so far using them proved to be a maintenance nightmare as lots of packaging tools have bugs around them, some of them quite hard to identify.

For at least one project I did a compromise and added a minor tests module inside the package while keeping the test suite outside. I only included the small bits that were needed for testing the plugins. I may had being lucky as this was not hard to to in my case.

Patch-set: 9