Update patch set 6
Patch Set 6:

(1 comment)

I tried implementing the whitelist (having several allowed_origins) but as far as I understood the specification it is not possible to have a list of origins it the Access-Control-Allow-Origin header or return multiple times the header. The only way to do that would be to check the origin on the fly and verify that it matches the whitelist. Not sure we want to do that, what do you think ?

Patch-set: 6