Lock/unlock nodes on executor server
Currently, the nodes are locked in the scheduler/pipeline manager before
the actual build is created in the executor client. When the nodes are
locked, the corresponding NodeRequest is also deleted.

With this change, the executor will lock the nodes before actually
starting the build and unlock them when the build is completed. To make
this work, we must also keep the NodeRequest active until the build is
completed.

This also moves the autohold processing from the scheduler to the
executor server. To make this work, the executor now also determines
build attempts and sets the RETRY_LIMIT result if necessary.

Change-Id: I7392ce47e84dcfb8079c16e34e0ed2062ebf4136