Support credentials supplied by nodepool
This adds support for SSH and WinRM keys/certs from nodepool.

This also uses the secret passing facility in the bubblewrap driver
(which supplied secrets on a pipe rather than writing them to the
filesystem) for not only certs/keys from nodepool but also the
existing executor-supplied winrm files.  This is slightly more
secure (in that they are never written to disk, and only appear
in the work directory of jobs that need them), but it does mean
that bubblewrap is required for the setup and cleanup ansible
runs.  This means there are no cases where we run un-wrapped ansible
so the "wrapped" parameter from runAnsible is removed.

Depends-On: https://review.opendev.org/774014
Change-Id: I6314e431a80640a630a2edc2a5824c6341176b9e