Make ZooKeeper mandatory for Scheduler
This makes the ZooKeeper connection mandatory on the Scheduler.

Currently, the there is only one place where the ZooKeeper connection is
not needed when a Scheduler is instantiated, which is the config
validation. To make this work without providing a Scheduler instance,
those methods where changed to be classmethods and all relevant
parameters are provided directly rather than taking them from the active
Scheduler instance.

Making Zookeeper a mandatory parameter in the Scheduler's constructor
helps to avoid race conditions or missing references.

Change-Id: I87f652d4c723d4f745b0120ff41905bc120d1846