Make ConnectionRegistry mandatory for Scheduler
So far the connection registry was added after the Scheduler was
instantiated.

This was useful to instantiate a Scheduler without connections for the
config validation. As this can now be done without instantiating a
scheduler, we can make the ConnectionRegistry mandatory to simplify the
Scheduler instantiation.

This also removes the "testonly" constructor parameter in Scheduler as
this became obsolete.

Change-Id: I7dc7d14dc43267365ae669e115a481f3b917451c