Use a full regex match on job implied branches
Every branch matcher is a regex.  Implied branch matchers are regexes
whose content is simply the name of the branch.  Therefore, an implied
branch matcher for the "dev" branch looks like "dev".  Interpreted as
a regex, this will additionally match the "develop" branch.

We already have code to handle that case for implied branch matchers
for project-templates.  They use the ImpliedBranchMatcher class which
uses the re.fullmatch method.  This change applies the same class to
implied branch matches for jobs as well.

Change-Id: I8662dd895fbf6a2ec41c1221ec4e3a99a01003cd