Improve typings in context of builds via ZooKeeper
This will improve the navigation and overall readability of the code.
It's kept as a separate change but mainly touches the parts that are
somehow relevant for the builds via ZooKeeper change that follows later
on.

One thing we should keep in mind: The ConnectionRegistry and the
Scheduler should always be imported conditionally when only used for
type checking:

from typing import TYPE_CHECKING

if TYPE_CHECKING:
    from zuul.lib.connections import ConnectionRegistry
    from zuul.scheduler import Scheduler

As those files by themselves do a lot of imports (they kind of have
access to everything), this will otherwise increase the loading time
for the imports. Which would then also increase the duration of the test
suite.

There are a few drawbacks in here where we need additional NoneType
checks for Optional variables, but a lot of those can be removed later
on when those attributes are mandatory (e.g. the ConnectionRegistry on
the Scheduler and the BuildSet and a Build).

Since Python 3.6 is now the minimum requirement PEP 526 [1] is used here.

[1]: https://www.python.org/dev/peps/pep-0526/

Change-Id: Iac23946dc61496b7199aca1719a23069c4f4e5de