DNM: Reduce number of jobs for SOS development
While developing the scale-out-scheduler we often have to rebase and
push a large stack of changes. This has quite a heavy impact on the
resources used by opendev Zuul.

Most of the time it's sufficient for us to have only the basic Python
tests running until we have finalized/polished the changes.

We can use this commit to reduce the number of tests for future
following commits and reduce the load on opendev Zuul.

Change-Id: I222ed7bfe3b850fa80d593fdacdf6a08aaff2fac