Update patch set 5
Patch Set 5: Code-Review-1

Okay, so I do see links added to the task summary view on a failure like you mentioned in the commit message. Looking at the summary for build 8481c857f536484e863e57b1e38b3df2 in particular though, it's finding things it thinks are URLs throughout, like setup.py and code=f.read which is a bit distracting. On the other current failure though, cbec8ae4e2774ddd8b6a8694faf3205f, I do indeed see an actual URL get linked ( https://www.docker.com/increase-rate-limit ), but the other string it links is docker.io/opendevorg/python-builder which is a 404. Perhaps it would be better to configure Linkify to only add hyperlinks to strings with a protocol:// prefix?

Why I originally didn't see hyperlinks getting added was that I was checking the logs and console views. I suppose BuildOutput.jsx is only for the Task Summary view? This is the particular detail I was hoping to see mentioned in the commit message (it claims to link the "build output" but it was not clear to me that was only the output in the summary and not elsewhere in the interface).

I'm not really all that familiar with Javascript, so I'm not entirely sure how to suggest altering the existing change, but I'm a soft -1 on at least makig it a little more strict on string matches and not just assuming anything with a . in it is a URL.

Patch-set: 5
Reviewer: Gerrit User 5263 <5263@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1