Update patch set 7
Patch Set 7: Code-Review-1

We should definitely handle this case, but I'm concerned with this approach, we're basically just guessing that HEAD^ is better than HEAD.  We don't know that's true, or that there even is a HEAD^.

As discussed in IRC, we could either delete the repo if the reset didn't work, or have Zuul write a commit that removes the .gitmodules file (it's not enough to simply remove it from the working tree -- while that worked in my simple constructed test case with a new empty repo, Albin found that it did not work in his copy of a production repo).  If we go with the second approach, we should make sure that all the code paths that call _git_fetch() then do something with the repo which would eliminate the fixup commit.  That's probably the case, but we should make sure we understand it fully.

Patch-set: 7
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1