Update patch set 6
Patch Set 6:

One concern was raised in IRC so far:

http://eavesdrop.openstack.org/irclogs/%23zuul/%23zuul.2021-02-12.log.html#t2021-02-12T23:17:56

In short, when we initially merged a change implementing parsing for ANSI color escape sequences in job output, the performance hit in the log browser was such that it severely increased the time required to display very large logfiles. If we add hyperlinking in the summary view, we'll probably want to extend it to the console and log browser views for consistency. In deciding to merge this change, we should perform at least some cursory performance tests of a similar implementation for the log  browser with some very large job logs.

A good example to try with is https://zuul.opendev.org/t/openstack/build/55c8bb6b766d4ced8515f3d7c505b572/log/controller/logs/screen-q-l3.txt or the same logfile from similar successful builds of the neutron-tempest-slow-py3 job for the openstack tenant on zuul.opendev.org, as it seems to regularly be in the vicinity of 25MB (and does have https:// URLs appearing for documentation URLs embedded in a couple of deprecation warnings around 10% of the way through the file).

Patch-set: 6
Reviewer: Gerrit User 5263 <5263@4a232e18-c5a9-48ee-94c0-e04e7cca6543>