Update patch set 6
Patch Set 6:

I understand the concern regarding the performance hit for large data sets, and the desire to be consistent across the Logs and Console tabs (especially since 2 reviewers of this change intuitively expected to find the impact of this change in those tabs).
But, should that be holding this back for Task Summary?

The motivation here was to address a common case (at least for our Zuul instance) where the last few lines of output in a failed task contain something like "See more information about the result: <URL>".
This is where an un-engaged developer, who doesn't really understand/use any of the other tabs (WTH is Ansible?), sighs and manually copies the URL to their browser.

As mentioned in the IRC log, Task Summary truncates to a small number of lines for each failed task, and to me the performance hit there is imperceptible.

Patch-set: 6