Update patch set 6
Patch Set 6: Code-Review-1

(1 comment)

I'm concerned with the performance, even with the limited size in the summary view.

Does using a stricter regex ("^https?://") improve the speed?  (See my inline comment.)

We seem to be getting good performance (I admit, I don't have numbers) using JS regex parsing for the severity filtering in the log view.  If the above idea doesn't pan out, maybe we can just try our own regex filtering.

I would like to eventually have ansi rendering too, and if the easiest way to get this done is to bundle both together, that's fine.   But doing so may take longer (it's necessarily going to be slower, and I think it's more important to be consistent with ansi on all the tabs than it necessarily is with links, and maybe we have to develop some kind of size limit cutout).  In short, it may take longer to get to a conclusion there, even if the result is better.  Let's consider whether we can make smaller gains in a shorter time before deciding to bundle them (we can always "upgrade" to a bundled solution later; it will be invisible to users).

Patch-set: 6
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1