Update patch set 7
Patch Set 7:

I took Tristan's benchmark change and then made an attempt at a really simple linkification based on regex replacement, just to see how that approach compares.  It runs the benchmark in 33ms on my workstation, comparable to the status quo (and Tristan's numbers as well).

I used the react-string-replace library because it's the first thing I found that could handle the jsx replacement.  It's very old and hasn't been updated in a while.  Maybe that's fine, but we should see if there's a better state of the art.

What I wrote also doesn't correctly handle the subgroup in the match.  But I think it's good enough to establish a baseline for performance.  Maybe we should go with that approach?  Feel free to crib anything useful from the quick hack I wrote, or throw it out if there's something better.  I've pushed it up as a followup to this change for reference: https://review.opendev.org/c/zuul/zuul/+/775506

Patch-set: 7
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>