Update patch set 1
Patch Set 1:

I think there is an opportunity in having zuul build images to fill a use case that is not currently covered by zuul/nodepool.

It would be really useful to have a job that builds an image, and then subsequent jobs consume that image to run integration/validation tests on it. Assuming they pass then it is published to nodepool or whatever.

This would be very useful for those building distros or appliances.

(This could also be done in a similar way to container images for consistency).

Patch-set: 1
CC: Gerrit User 7069 <7069@4a232e18-c5a9-48ee-94c0-e04e7cca6543>