Move setupZK() helper function to BaseTestCase class
As we are going to move initialization of the ZooKeeper connection to
the server classes rather than the cmd classes, this should help to
provide the necessary parameters to the various components in the test
cases.

This also removes some duplicate code as the setupZK() functionality can
now also  be used by other TestClasses which don't inherit from
ZuulTestCase.

Change-Id: Iee64b7173ceb78967e48dcd6f51bc998ce9469c5