Implement ABC for caching changes in Zookeeper
In order to make updates atomic the change data is stored separate from
the cache entry. The data uses a random UUID znode that is then
referenced from the actual cache entry.

The change data is immutable, which means that an update of a cached
item will result in a new data node. The cache entry will then be
changed to reference the new data.

This approach also allows us to check if a given change is up-to-date by
comparing the referenced UUID in Zookeeper with the one in the local
cache without loading the whole change data.

The change data is stored in the following Zookeeper path:

    /zuul/cache/connection/<connection-name>/data/<uuid>

The cache entries that reference the change data use the following path:

    /zuul/cache/connection/<connection-name>/cache/<key>

Data nodes will not be directly removed when an entry is removed or
updated in order to prevent race conditions with multiple consumers of
the cache. The stale data nodes will be instead cleaned up in the
cache's cleanup() method. This expected to happen periodically.

TODO:
* interface for cleaning up entries older than X

Change-Id: I9a020b65c2c2e2326d5eb8bbd16270de4f4c1eab