Switch to ZooKeeper backed build result events
This puts the build result events (BuildStarted, BuildStatus,
BuildPaused and BuildCompleted) into ZooKeeper.

This change moves most of the result event logic into the scheduler
class. Previously, the executor client did a lot before forwarding the
events to the scheduler. Now, the scheduler picks up the events from the
result event queues and does everything by its own before notifying the
pipeline manager.

To make this work, noop builds now get an entry in the local self.builds
list in the executor client. Like before, they are started and completed
immediately (by putting the corresponding events in the result event
queue) without acquiring a compute node.

To properly deal with the pending build results from tenants/pipelines
that were removed during a reconfiguration, we have to put some cleanup
logic into the Scheduler. This is necessary as the result events of
those builds cannot be processed anymore as the tenant/pipeline is no
longer part of the current abide.

The builds itself are still executed via Gearman, but the results aren't
looked up from the Gearman job anymore, but from the result queues in
ZooKeeper.

Change-Id: I067c38cec8ac14f3a3bb6f607ede943c88ce493d