Abort build if update raises BrokenProcessPool
On occasion (e.g. due to an OOM condition), BrokenProcessPool can
occur within the ExecutorServer._innerUpdateLoop() method. If it
happens in AnsibleJob.execute() we send an abort so the build can be
retried automatically, so let's do the same for _innerUpdateLoop()
rather than just completing it in an ERROR state.

Change-Id: I33c4737876a49bb7ac1a54a7783e3a6389ba35fe