Emit config warnings if shared queues on pipelines
In I14533ae57cfe688b55c26a1e17ab38e133180b28 we moved the shared
queue definition from the pipeline to the project.  Now that users
have had an opportunity to move to using the new syntax, begin
emitting warning messages if the old syntax is used.

If a user uploads a change with the old syntax, Zuul will leave a
warning message on the change comment.  Additionally, existing
occurences of the old syntax will appear in the web UI's config
errors listing.  They are currenty visually un-differentiated from
errors (aside from their text), though this change adds metadata
to the JSON representation that could be used in a followup change
to visually indicate they are warnings.

Change-Id: I4d24de80180ab8416bb6972ceb17a72891dca653