Update zoned executor stats
This gives us 3 sets of executor stats: zoned, unzoned, and all.

The zoned and unzoned stats are moved to a hierarchy that avoids
name collisions.

The only way to detect whether an executor in a zone is online
is with a function with its zone name registered.  Since the only
function currently registered with a zone name is "execute:execute"
and that is unregistered when an executor is online but not accepting,
we need to add a new dummy function in order to do this accounting.

This also updates the docs and adds a release note about the (minor)
change in stats meaning.

Change-Id: Ie28963426024f2d54275426794549f31ace9d998