Update patch set 2
Patch Set 2: Code-Review-1

In IRC we discussed the idea that "all" is going to be complicated by change 673840 (allow unzoned option), in that we will have executors that can handle zoned and unzoned jobs.  In that case, the "all" category becomes more complicated, and in a mixed zoned/unzoned system, it can actually be misleading regardless.  It doesn't matter if you have 5 executors running if they are all zoned and you need an unzoned one.

So instead, let's keep what this change does with zoned and unzoned, but make the top-level hierarchy be a copy of unzoned, announce that it's deprecated and existing dashboards should update to "unzoned" keys, and then remove the top level in 5.0.

Patch-set: 2
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1