DNM Deactivate statsd collection for online executors and running builds
The current implementation of the stats collection requires Gearman. We
could utilize ZK for this purpose, but this would require some changes
in the executor and scheduler which are only necessary for the stats
collection, but for nothing else.

Thus, let's deactivate this for now and discuss the topic first, before
changing the implementation.

Change-Id: Ib73441274c0d8457e13aac1aac78c4335099bd4e