Change Description of DB in Components Page on Zuul Document
At present, if we read components page[1] of Zuul doc, database
seems to be required only when SQL connection is specified in zuul.conf.
However, database is required even if SQL connection is not specified
in zuul.conf.

To avoid this confusion, this commit modifies components page of doc.

[1] https://zuul-ci.org/docs/zuul/discussion/components.html

Story: 2008725
Task: 42069
Change-Id: Id1d68bd5cff43e9479dab851018799691c8e3f57