Fix trigger event forwarding bug
The change to move trigger events into ZK had an error.  It tries
to determine if it needs to forward each trigger event to each
pipeline in each tenant.  To do that, it re-uses the pipeline
management functions to determine if an event might be relevant to
a change in that pipeline (for example, it is an abandon event).
Those comparisons have not been used to compare items of different
types before (changes to tags, for example), and raise exceptions
in those cases.

To correct this, improve the equality comparisons for those types.

Additionally, when this exception was raised, it caused any further
processing of the event to abort, so any pipelines or tenants after
the tenant-pipeline which caused the error would not receive the
event.  To make the system more robust so that an error related to
one tenant does not affect others, we add an extra exception handler
to log those errors and proceed to the next tenant.

Change-Id: I9b62b83d51f66bee85cdb968d24bd5367e62e35d