Fix bug w/ None event in Gitub event forwarder
Change Ie54fc16488ab8cbc15f97d003f36c12b8a648ed4 introduced a bug in the
Github event forwarder thread that is triggered whenever the
GithubEventProcessor doesn't produce an event.

In case the event was None the GithubEventForwarder stopped instead of
continuing with the next future in the queue.

Change-Id: Ibd76c1a787f5b8bd186bcb271bc526fe5e5ddade