Don't refresh change when enqueuing an dequeue event
We only fetch the change object in order to verify that it's a real
change and to compare it to the queue item.  We don't need to have
up-to-date information about the change for that.

On the enqueue side, we do have further use of the change object
in that it is handed off to the pipeline manager for enqueing, so
we should refresh the change, just as we do when processing trigger
events.

Change-Id: Ibe835d22945c8c42c37ad97baaf0500b5f809c5d