Use ephemeral nodes for management result events
We are concerned about the case where an external client submits
a management event which requires a result, but then disconnects
from ZK before retrieving the result.  Currently, the client is
responsible for deleting the result node once it is populated; in
this case, the result node could leak.

To avoid this, let's have the client pre-create the result node as
an ephemeral node, and then the processor of the event can write to
that node (if it exists, or log a warning if it doesn't).  If the
client happens to disconnect before the result is written, the node
will still be cleaned up.

An ephemeral node makes sense in this case since no other component
of the system needs the results of the management event other than
the originating client.

Change-Id: I420235ca748c4c28b40aaf4453e7805f73dfd5ba