Fix race condition related to out-of-band acks
The Github connection consumes events from the Zookeeper event queue and
dispatches them to a thread pool for processing. Events are then acked
from a different thread.

This can lead to a situation where a consumer started iterating over the
events while another event that was not yet yielded is acked. In that
case the iteration will abort with a NoNodeError exception when it
trieds to return this event.

In order to fix that, we can simply ignore the NoNodeError in the
__iter__ method and continue with the next event.

Change-Id: I6f6d7a02ca8358c26ca1afa1d1bfaf9341ca2c3f