Fix ZK-related race condition in github driver
This change is an alternative to I6f6d7a02ca8358c26ca1afa1d1bfaf9341ca2c3f.

At a high level, our plan for ZooKeeper queues is that anyone can append to
a queue without a lock, but a queue would only have one processor at a time
which is able to remove items from the queue.  In the case of the github
driver, that is generally true, except that we have two threads doing that
work; the first processes queue items, and the second actually removes them.
The two threads in some respects act like two simultaneous "owners" of the
queue.  This cas cause a problem within a single process (in that the second
thread can change the queue out from under the first) but also potentially
in multiple processes if we don't drain events from the second thread before
stopping the first (the second thread could continue to remove events
out from under a *different* process in that case).

To address both of these, we move all of the event handling into a single
thread.  This maintains our "single-owner" paradigm, while also maintaining
the benefits of the concurrent processing.  We start each cycle of the
dispatch loop by reading events from the queue and beginning new pre-processing
of those events, and we finish each cycle by collecting and acking any events
whose pre-processing is complete.

The events_in_progress set no longer has to contend with a second thread, so
we don't need to make a copy of it anymore.

Change-Id: I42f49a60ac3494c39f62cf63e35b78fb5fc0d804