Switch to ZooKeeper backed merge result events
This puts the merger result events (MergeCompleted, FilesChanged) into
ZooKeeper.

This is the first step to put the merge jobs into ZooKeeper, similar to
the builds.

This doesn't change the logic how the result events are processed. The
main change is done in the scheduler callbacks (onMergeCompleted,
onFilesChanged) which now put a serializable merge result event into the
ZooKeeper result event queue rather than the local queue. The event
handler methods are adapted to work with those new events.

As we buildset is not serializable in its current state, we provide the
buildset UUID to the events and look up the corresponding buildset in
the event processing methods based on the provided pipeline and queue.

Change-Id: I033cf27bc8035afbd743e37292da37fde6d0e0b8