Update branches after fast forward merges
Previously Zuul would check two things to determine if a repo in the
merger needs updating:

  1) Does the ref exist in the merger repo? if not then update
  2) Does the rev referred to by ref in the upstream repo exist in the
     merger repo? if not then update

Unfortunately, this is not a complete set of checks as fast forward
mergers of changes can undermine the existing checks. Consider a merger
that handed refs/changes/34/1234 previously. It will have the rev for
that change present in the repo but ref does not refer to it. Now some
time later the change is merged via a fast forward to ref. Now the
upstream ref refers to rev, but when the merger checks if it should
update the local repo to pick up this update it decides this isn't
necessary.

This happens because the local repo already has the ref
(master/main/etc) and it has the rev (from a previous pre merge testing
attempt).

We fix this by adding a third check:

  3) Does the merger version of ref refer to rev? if not then update

Change-Id: I5eeab42b4cca5dd415a2b5d9438b32b18c45562f