Execute merge jobs via ZooKeeper
This change implements a ZooKeeper MergerApi that stores merge requests
in ZooKeeper.

The merge requests are stored in the path:
    /zuul/merge-requests/<uuid>

The merger client puts new merge requests in ZooKeeper. The merger
server listens for those via a children watch on the merge-requests
path.

Instead of notifying the client first, the merger server now puts the
merge result directly in the corresponding result queue in ZooKeeper.
Afterwards, the merge server deletes the merge request.

In case the caller wants to wait for the result synchronously (e.g.
during config loading), the merger will provide a MergeResultEventFuture
instead which contains the path to the result in ZooKeeper.

Change-Id: I767c0b4c5473b2948487c3ae5bbc612c25a2a24a