Update patch set 1
Patch Set 1: Code-Review-1

We didn't find any evidence that this is currently causing a problem.  And this code should not cause any leaks -- the only thing that makes it out of this exception handler is a string (unlike the previous incidence where we passed exception objects outside the handler).  Without evidence that it's a problem, this seems like a bit of voo-doo.  I'm willing to accept that there could be something fundamentally strange about frame objects that necessitate this, but unless we either see evidence of that, or someone feels really strongly about this, I'd just as soon not merge it since it adds complexity that we can't adequately explain.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1