Periodically clean up leaked semaphores
If there happens to be a bug in Zuul, or if the scheduler crashes,
then we may leak semaphores, which, now that they are held in ZK,
is a permanent situation.  To avoid the necessity of operators
manually editing or removing ZK nodes, introduce a periodic semaphore
cleanup.

This identifies any semaphores which are held by jobs which are not
running and therefore may be removed.  It runs once an hour, outside
the main scheduler thread.  It also runs immediately after the first
reconfiguration, so that if a scheduler crashes and is restarted,
leaked semaphores are cleaned up quickly.

The general algorithm should be safe with multiple schedulers, though
the details about how we should traverse the entire abide are unclear
at this point, so this may need to change.  But it is as forward-
compatible as I can make it right now.  We may also decide that once
all running builds are in ZK, we no longer need this.  Regardless,
it is needed now, and should be adaptible to future situtaions if we
want it.

Change-Id: I023ddabb13047a1ed95a26a8f5793de36817c191