Update patch set 1
Patch Set 1: Code-Review-1

The purpose of the isUpdateNeeded function is to check if we need to fetch such that we are able to restore the repo state that gets passed in. So I wonder if the problem is something else. Now thinking again at the change that added the ref checks [1] (I don't think that those should be needed either) I think the real problem might be somewhere else and we're papering over a different issue here.

A second look at the log in the commit message of that change [2] shows that the executor is skipping the update, then merging and then restoring the repo state which is weird since it is expected to restore the repo state before it does anything else with the repo.

Before merging this change we should have a deeper look into this sequence since I believe when we fix the sequence this change is not needed at all (and the ref checking as well) and just will make the 'fetch only if necessary' much less effective.

[1] https://review.opendev.org/c/zuul/zuul/+/678895
[2] https://review.opendev.org/c/zuul/zuul/+/678895/3//COMMIT_MSG

Patch-set: 1
Reviewer: Gerrit User 16068 <16068@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1