Update patch set 1
Patch Set 1:

I think I've found the missing place where we need to restore the repo state [1]. With global repo state all repos that weren't part of the dependency chain are now part of the repo state which causes the skip. However after the update call (that skips the update if the rev is already part of the repo state) we don't seem to restore that.

[1] https://opendev.org/zuul/zuul/src/commit/991d8280ac54d22a8cd3ff545d3a5e9a2df76c4b/zuul/executor/server.py#L1053

Patch-set: 1
Reviewer: Gerrit User 16068 <16068@4a232e18-c5a9-48ee-94c0-e04e7cca6543>