Keep jobgraphs frozen across reconfiguration
This removes test cases which are no longer be relevant.

Many of these were testing various mutations across reconfigurations,
but with job graphs frozen, about the only thing that we expect to
change now is when a pipeline, project, or tenant is deleted.  Test
cases are modified or added to test these.

It appears even the current code may have some bugs related to deleting
pipelines and tenants.  The improved testing in this change highlighted
that.  The scheduler is updated to ensure that it cancels all jobs on
pipelines or tenants that are removed from a running configuration.  This
should ensure we don't leak nodes or semaphores.

Change-Id: I2e4bd2fb9222b49cb10661d28d4c52a3c994ba62